requirements.txt
# Основные зависимости
streamlit>=1.28.0
gigachat>=0.1.9
chromadb>=0.4.15
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.0
requests>=2.31.0
pymupdf>=1.23.0
python-docx>=1.1.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
nltk>=3.8.0
lxml>=4.9.0

# Разработка и тестирование
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
pylint>=2.17.0
pre-commit>=3.0.0